Updated the OAuth2 sample application.
