+++ samples/oauth2/app/src/main/java/com/apigee/oauth2/grantTypeActivities/ClientCredentialsGrantTypeActivity.java
@@ -1,7 +1,6 @@
-package com.apigee.oauth2;
+package com.apigee.oauth2.grantTypeActivities;
 
 import android.app.Activity;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
@@ -9,16 +8,12 @@ import android.widget.Button;
 import android.widget.TextView;
 
 import com.apigee.google.api.client.auth.oauth2.TokenResponse;
+import com.apigee.oauth2.Client;
+import com.apigee.oauth2.Constants;
+import com.apigee.oauth2.R;
+import com.apigee.oauth2.asyncTasks.PaloAltoTemperatureDownloadTask;
 import com.apigee.sdk.data.client.ApigeeDataClient;
 import com.apigee.sdk.data.client.callbacks.OAuth2ResponseCallback;
-
-import org.json.JSONObject;
-
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
 
 
 public class ClientCredentialsGrantTypeActivity extends Activity {
@@ -75,8 +70,8 @@ public class ClientCredentialsGrantTypeActivity extends Activity {
 
     public void clearData() {
         this.tokenResponse = null;
-        this.accessTokenTextView.setText("N/A");
-        this.paloAltoTemperatureTextView.setText("N/A");
+        this.accessTokenTextView.setText(getString(R.string.not_available));
+        this.paloAltoTemperatureTextView.setText(getString(R.string.not_available));
     }
 
     public void storeToken() {
@@ -121,84 +116,11 @@ public class ClientCredentialsGrantTypeActivity extends Activity {
 
     public void getPaloAltoTemperature() {
         String accessTokenString = this.accessTokenTextView.getText().toString();
-        if( !accessTokenString.equalsIgnoreCase("N/A") ) {
+        if( !accessTokenString.equalsIgnoreCase(getString(R.string.not_available)) ) {
             String temperatureURL = String.format(Constants.kApigeeClientCredentialsWeatherInfoURLFormat,Constants.ORG_ID);
-            new PaloAltoTemperatureDownloadTask(temperatureURL,accessTokenString).execute();
-        }
-    }
-
-    private class PaloAltoTemperatureDownloadTask extends AsyncTask<Void, Void, String> {
-
-        private String temperatureURL;
-        private String accessToken;
-
-        public PaloAltoTemperatureDownloadTask(String temperatureURL, String accessToken) {
-            this.temperatureURL = temperatureURL;
-            this.accessToken = accessToken;
-        }
-
-        @Override
-        protected String doInBackground(Void... params) {
-
-            String temperatureString = null;
-
-            HttpURLConnection urlConnection = null;
-            InputStream inputStream = null;
-
-            try {
-                URL url = new URL(temperatureURL);
-                urlConnection = (HttpURLConnection) url.openConnection();
-
-                urlConnection.setRequestMethod("GET");
-                urlConnection.setUseCaches(false);
-
-                if  ( accessToken != null && accessToken.length() > 0 ) {
-                    String authStr = "Bearer " + accessToken;
-                    urlConnection.setRequestProperty("Authorization", authStr);
-                }
-
-                urlConnection.setDoInput(true);
-
-                inputStream = urlConnection.getInputStream();
-
-                if( inputStream != null ) {
-                    BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
-                    StringBuilder sb = new StringBuilder();
-                    String line;
-
-                    while ((line = reader.readLine()) != null) {
-                        sb.append(line);
-                        sb.append('\n');
-                    }
-
-                    JSONObject jsonObject = new JSONObject(sb.toString());
-                    JSONObject weatherAttributes = jsonObject.getJSONObject("rss")
-                            .getJSONObject("channel")
-                            .getJSONObject("item")
-                            .getJSONObject("yweather:condition")
-                            .getJSONObject("#attrs");
-                    temperatureString = weatherAttributes.getString("@temp") + "" + (char)186;
-                }
-            } catch ( Exception exception ) {
-                Log.d("Exception getting temperature: ",exception.getLocalizedMessage());
-            } finally {
-                try {
-                    if( inputStream != null ) {
-                        inputStream.close();
-                    }
-                    if( urlConnection != null ) {
-                        urlConnection.disconnect();
-                    }
-                } catch(Exception ignored) {
-                }
-            }
-            return temperatureString;
-        }
-
-        protected void onPostExecute(String temperature){
-            if( temperature != null && temperature.length() > 0 ) {
-                ClientCredentialsGrantTypeActivity.this.paloAltoTemperatureTextView.setText(temperature);
-            }
+            new PaloAltoTemperatureDownloadTask(this,this.paloAltoTemperatureTextView,temperatureURL,accessTokenString).execute();
+        } else {
+            Client.showAlert(this, "Error getting temperature", "Access token is not valid.");
         }
     }
 }
