+++ samples/oauth2/app/src/main/java/com/apigee/oauth2/grantTypeActivities/PasswordGrantTypeActivity.java
@@ -1,25 +1,18 @@
-package com.apigee.oauth2;
+package com.apigee.oauth2.grantTypeActivities;
 
 import android.app.Activity;
-import android.os.AsyncTask;
 import android.os.Bundle;
-import android.util.Log;
 import android.view.View;
 import android.widget.Button;
 import android.widget.TextView;
 
 import com.apigee.google.api.client.auth.oauth2.TokenResponse;
+import com.apigee.oauth2.Client;
+import com.apigee.oauth2.Constants;
+import com.apigee.oauth2.R;
+import com.apigee.oauth2.asyncTasks.ApigeeEmailAddressDownloadTask;
 import com.apigee.sdk.data.client.ApigeeDataClient;
 import com.apigee.sdk.data.client.callbacks.OAuth2ResponseCallback;
-
-import org.json.JSONArray;
-import org.json.JSONObject;
-
-import java.io.BufferedReader;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
 
 
 public class PasswordGrantTypeActivity extends Activity {
@@ -75,8 +68,9 @@ public class PasswordGrantTypeActivity extends Activity {
     }
 
     public void clearData() {
-        this.accessTokenTextView.setText("N/A");
-        this.emailAddressTextView.setText("N/A");
+        this.tokenResponse = null;
+        this.accessTokenTextView.setText(getString(R.string.not_available));
+        this.emailAddressTextView.setText(getString(R.string.not_available));
     }
 
     public void storeToken() {
@@ -104,98 +98,27 @@ public class PasswordGrantTypeActivity extends Activity {
         String tokenURL = String.format(Constants.kApigeePasswordGrantTokenURLFormat,Constants.ORG_ID,Constants.APP_ID);
 
         final ApigeeDataClient dataClient = Client.sharedClient().dataClient();
-        dataClient.oauth2AccessTokenAsync(tokenURL,Constants.kApigeePasswordGrantUsername,Constants.kApigeePasswordGrantPassword,Constants.kApigeeClientCredentialsClientID,new OAuth2ResponseCallback() {
+        dataClient.oauth2AccessTokenAsync(tokenURL, Constants.kApigeePasswordGrantUsername, Constants.kApigeePasswordGrantPassword, Constants.kApigeeClientCredentialsClientID, new OAuth2ResponseCallback() {
             @Override
             public void onResponse(TokenResponse response) {
-                if( response != null && response.getAccessToken() != null ) {
+                if (response != null && response.getAccessToken() != null) {
                     PasswordGrantTypeActivity.this.tokenResponse = response;
                     PasswordGrantTypeActivity.this.accessTokenTextView.setText(response.getAccessToken());
                 }
             }
             @Override
             public void onException(Exception e) {
-
             }
         });
     }
 
     public void getEmailAddress() {
         String accessTokenString = this.accessTokenTextView.getText().toString();
-        if( !accessTokenString.equalsIgnoreCase("N/A") ) {
+        if( !accessTokenString.equalsIgnoreCase(getString(R.string.not_available)) ) {
             String emailAddressURL = String.format(Constants.kApigeePasswordGrantUserInfoURLFormat,Constants.ORG_ID,Constants.APP_ID,Constants.kApigeePasswordGrantUsername);
-            new ApigeeEmailAddressDownloadTask(emailAddressURL,accessTokenString).execute();
-        }
-    }
-
-    private class ApigeeEmailAddressDownloadTask extends AsyncTask<Void, Void, String> {
-
-        private String emailAddressURL;
-        private String accessToken;
-
-        public ApigeeEmailAddressDownloadTask(String emailAddressURL, String accessToken) {
-            this.emailAddressURL = emailAddressURL;
-            this.accessToken = accessToken;
-        }
-
-        @Override
-        protected String doInBackground(Void... params) {
-
-            String emailAddress = null;
-
-            HttpURLConnection urlConnection = null;
-            InputStream inputStream = null;
-
-            try {
-                URL url = new URL(emailAddressURL);
-                urlConnection = (HttpURLConnection) url.openConnection();
-
-                urlConnection.setRequestMethod("GET");
-                urlConnection.setUseCaches(false);
-
-                if  ( accessToken != null && accessToken.length() > 0 ) {
-                    String authStr = "Bearer " + accessToken;
-                    urlConnection.setRequestProperty("Authorization", authStr);
-                }
-
-                urlConnection.setDoInput(true);
-
-                inputStream = urlConnection.getInputStream();
-
-                if( inputStream != null ) {
-                    BufferedReader reader = new BufferedReader(new InputStreamReader(inputStream));
-                    StringBuilder sb = new StringBuilder();
-                    String line;
-
-                    while ((line = reader.readLine()) != null) {
-                        sb.append(line);
-                        sb.append('\n');
-                    }
-
-                    JSONObject jsonObject = new JSONObject(sb.toString());
-                    JSONArray entitiesArray = jsonObject.getJSONArray("entities");
-                    JSONObject entityObject = (JSONObject) entitiesArray.get(0);
-                    emailAddress = entityObject.getString("email");
-                }
-            } catch ( Exception exception ) {
-                Log.d("Exception getting email address: ",exception.getLocalizedMessage());
-            } finally {
-                try {
-                    if( inputStream != null ) {
-                        inputStream.close();
-                    }
-                    if( urlConnection != null ) {
-                        urlConnection.disconnect();
-                    }
-                } catch(Exception ignored) {
-                }
-            }
-            return emailAddress;
-        }
-
-        protected void onPostExecute(String emailAddress){
-            if( emailAddress != null && emailAddress.length() > 0 ) {
-                PasswordGrantTypeActivity.this.emailAddressTextView.setText(emailAddress);
-            }
+            new ApigeeEmailAddressDownloadTask(this,this.emailAddressTextView,emailAddressURL,accessTokenString).execute();
+        } else {
+            Client.showAlert(this, "Error getting email address.", "Access token is not valid.");
         }
     }
 }
