+++ samples/oauth2/app/src/main/res/layout/activity_authorization_code_grant_type.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.apigee.oauth2.ImplicitGrantTypeActivity"
+    tools:context="com.apigee.oauth2.grantTypeActivities.ImplicitGrantTypeActivity"
     android:id="@+id/AuthorizationCodeGrantTypeActivityView">
 
     <TextView
@@ -33,7 +33,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="N/A"
+        android:text="@string/not_available"
         android:id="@+id/acAccessTokenTextView"
         android:maxWidth="250dp"
         android:layout_below="@+id/acAccessTokenHeaderTextView"
@@ -44,7 +44,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Email Addres:"
+        android:text="Email Address:"
         android:id="@+id/acEmailAddressHeaderTextView"
         android:textStyle="bold"
         android:layout_below="@+id/acAccessTokenTextView"
@@ -54,7 +54,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="N/A"
+        android:text="@string/not_available"
         android:id="@+id/acEmailAddressTextView"
         android:layout_below="@+id/acEmailAddressHeaderTextView"
         android:layout_centerHorizontal="true"
