+++ samples/oauth2/app/src/main/res/layout/activity_client_credentials_grant_type.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.apigee.oauth2.ClientCredentialsGrantTypeActivity"
+    tools:context="com.apigee.oauth2.grantTypeActivities.ClientCredentialsGrantTypeActivity"
     android:id="@+id/clientCredsActivityView">
 
     <TextView
@@ -33,7 +33,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="N/A"
+        android:text="@string/not_available"
         android:id="@+id/ccAccessTokenTextView"
         android:layout_below="@+id/ccAccessTokenHeaderTextView"
         android:layout_centerHorizontal="true"
@@ -42,7 +42,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Temp In Palo Alto"
+        android:text="Temp In Palo Alto:"
         android:id="@+id/ccTempInPaloAltoHeaderTextView"
         android:layout_below="@+id/ccAccessTokenTextView"
         android:layout_centerHorizontal="true"
@@ -52,7 +52,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="N/A"
+        android:text="@string/not_available"
         android:id="@+id/ccTempInPaloAltoTextView"
         android:layout_below="@+id/ccTempInPaloAltoHeaderTextView"
         android:layout_centerHorizontal="true"
