+++ samples/oauth2/app/src/main/res/layout/activity_implicit_grant_type.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.apigee.oauth2.ImplicitGrantTypeActivity"
+    tools:context="com.apigee.oauth2.grantTypeActivities.ImplicitGrantTypeActivity"
     android:id="@+id/ImplicitGrantTypeActivityView"
     android:clickable="true">
 
@@ -34,7 +34,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="N/A"
+        android:text="@string/not_available"
         android:id="@+id/iiAccessTokenTextView"
         android:maxWidth="250dp"
         android:layout_below="@+id/iiAccessTokenHeaderTextView"
@@ -45,7 +45,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Email Addres:"
+        android:text="Email Address:"
         android:id="@+id/iiEmailAddressHeaderTextView"
         android:textStyle="bold"
         android:layout_below="@+id/iiAccessTokenTextView"
@@ -55,7 +55,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="N/A"
+        android:text="@string/not_available"
         android:id="@+id/iiEmailAddressTextView"
         android:layout_below="@+id/iiEmailAddressHeaderTextView"
         android:layout_centerHorizontal="true"
