+++ samples/oauth2/app/src/main/res/layout/activity_password_grant_type.xml
@@ -6,7 +6,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.apigee.oauth2.PasswordGrantTypeActivity">
+    tools:context="com.apigee.oauth2.grantTypeActivities.PasswordGrantTypeActivity">
 
     <TextView
         android:text="Password"
@@ -32,7 +32,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="N/A"
+        android:text="@string/not_available"
         android:id="@+id/pAccessTokenTextView"
         android:layout_below="@+id/pAccessTokenHeaderTextView"
         android:layout_centerHorizontal="true"
@@ -42,7 +42,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Email Addres"
+        android:text="Email Address:"
         android:id="@+id/pEmailHeaderTextView"
         android:textStyle="bold"
         android:layout_below="@+id/pAccessTokenTextView"
@@ -52,7 +52,7 @@
     <TextView
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="N/A"
+        android:text="@string/not_available"
         android:id="@+id/pEmailTextView"
         android:layout_below="@+id/pEmailHeaderTextView"
         android:layout_centerHorizontal="true"
