+++ samples/oauth2/app/src/main/res/menu/authorization_code_grant_type.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.apigee.oauth2.AuthorizationCodeGrantTypeActivity" >
+    tools:context="com.apigee.oauth2.grantTypeActivities.AuthorizationCodeGrantTypeActivity" >
     <item android:id="@+id/action_settings"
         android:title="@string/action_settings"
         android:orderInCategory="100"
