+++ samples/oauth2/app/src/main/res/menu/password_grant_type.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.apigee.oauth2.PasswordGrantTypeActivity" >
+    tools:context="com.apigee.oauth2.grantTypeActivities.PasswordGrantTypeActivity" >
     <item android:id="@+id/action_settings"
         android:title="@string/action_settings"
         android:orderInCategory="100"
