+++ samples/oauth2/app/src/main/res/values/strings.xml
@@ -4,6 +4,7 @@
     <string name="app_name">oauth2</string>
     <string name="hello_world">Hello world!</string>
     <string name="action_settings">Settings</string>
+    <string name="not_available">N/A</string>
     <string name="title_activity_password_grant_type">PasswordGrantTypeActivity</string>
     <string name="title_activity_client_credentials_grant_type">ClientCredentialsGrantTypeActivity</string>
     <string name="title_activity_implicit_grant_type">ImplicitGrantTypeActivity</string>
